d3b_cli_igor/requirements.txt
boto3
airspeed>=0.5.14
amazon_kclpy-ext==1.5.1
aws-sam-translator>=1.15.1
awscli>=1.14.18
boto>=2.49.0
boto3>=1.14.33                 #basic-lib
botocore>=1.12.13
cachetools>=3.1.1,<4.0.0
cbor2>=5.2.0
coverage==4.5.4
crontab>=0.22.6
cryptography<3.4
dnspython==1.16.0              #basic-lib
docopt>=0.6.2                  #basic-lib
elasticsearch>=7.0.0,<8.0.0
flake8>=3.6.0                  #extended-lib
flake8-quotes>=0.11.0          #extended-lib
flask>=1.0.2
flask-cors>=3.0.3,<3.1.0
flask_swagger==0.2.12
forbiddenfruit==0.1.3
jsondiff>=1.2.0
jsonpatch>=1.24,<2.0
jsonpath-rw>=1.4.0,<2.0.0
localstack-ext[full]>=0.12.5
localstack-ext>=0.12.5         #basic-lib
localstack-client>=1.12        #basic-lib
moto-ext[all]>=2.0.0
nose>=1.3.7
nose-timer>=0.7.5
psutil>=5.4.8,<6.0.0
pympler>=0.6
pyopenssl==17.5.0
python-coveralls>=2.9.1
pyyaml>=3.13,<=5.1             #basic-lib
Quart>=0.6.15
requests>=2.20.0               #basic-lib
requests-aws4auth==0.9
sasl>=0.2.1
six>=1.12.0                    #basic-lib
xmltodict>=0.11.0
